Implemented basic controller endpoints

+++ src/main/java/com/nab/toyrobot/model/Position.java
@@ -6,6 +6,6 @@ import lombok.Data;
 @Data
 @Builder
 public class Position {
-    private int x;
-    private int y;
+    private Long x;
+    private Long y;
 }
